functions/requirements.txt
click==8.1.3
cloudevents==1.2.0
colorama==0.4.4
deprecation==2.1.0
Flask==2.1.2
functions-framework==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
watchdog==1.0.2
Werkzeug==2.1.2
